+++ mkinit/__main__.py
@@ -13,8 +13,12 @@ def main():
 
         The following `__init__.py` variables modify autogeneration behavior:
 
-            `__submodules__` - indicates the list of submodules to be
-                introspected, if unspecified all submodules are introspected.
+            `__submodules__` (List[str] | Dict[str, List[str])) -
+                Indicates the list of submodules to be introspected, if
+                unspecified all submodules are introspected. Can be a list
+                of submodule names, or a dictionary mapping each submodule name
+                to a list of attribute names to expose. If the value is None,
+                then all attributes are exposed (or __all__) is respected).
 
             `__external__` - Specify external modules to expose the attributes of.